requirements.txt
evdev==1.4.0
Pillow==8.3.2
pynput==1.7.3
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
python-xlib==0.31
six==1.16.0
